+++ services/graphics-server/src/api/shapes.rs
@@ -126,7 +126,7 @@ impl Rectangle {
 
 impl Into<ClipRect> for Rectangle {
     fn into(self) -> ClipRect {
-        ClipRect::new(self.x0(), self.y0(), self.x1(), self.y1())
+        ClipRect::new(self.x0() as u32, self.y0() as u32, self.x1() as u32, self.y1() as u32)
     }
 }
